More logic fixes.  Now, VFAT devices can have folders double-clicked on to add items to the playlist.  Also, both folders and tracks can now be dragged and dropped to move files and folders around.  Seems to be pretty stable.  Hopefully I didn't break anything ipod-related :-)
svn path=/trunk/extragear/multimedia/amarok/; revision=537059

+++ src/mediadevice/vfat/vfatmediadevice.cpp
@@ -843,9 +843,10 @@ VfatMediaDevice::dirListerClear( const KURL &url )
 {
     DEBUG_BLOCK
     QString directory = url.path(-1);
-    debug() << "Removing: " << directory << endl;
+    debug() << "Removing url: " << directory << endl;
     VfatMediaFile *vmf = m_mfm[directory];
-    vmf->deleteAllChildren();
+    if( vmf )
+        vmf->deleteAllChildren();
 }
 
 void
@@ -853,9 +854,10 @@ VfatMediaDevice::dirListerDeleteItem( KFileItem *fileitem )
 {
     DEBUG_BLOCK
     QString filename = fileitem->url().path(-1);
-    debug() << "Removing file: " << filename << endl;
+    debug() << "Removing item: " << filename << endl;
     VfatMediaFile *vmf = m_mfm[filename];
-    vmf->deleteAll();
+    if( vmf )
+        vmf->deleteAll();
 }
 
 int
